refactor: 클래스, 폴더 이동
